Renamed all files to reflect the new model name
