+++ netbox_dns/api/serializers.py
@@ -1,7 +1,7 @@
-from .serializers_.contact import *
 from .serializers_.nameserver import *
 from .serializers_.record import *
 from .serializers_.registrar import *
+from .serializers_.registration_contact import *
 from .serializers_.view import *
 from .serializers_.zone import *
 from .serializers_.zone_template import *
